requirements.txt
cvxopt
Cython==0.29.21
importlib_metadata==4.11.3
matplotlib
ncg_optimizer
numpy==1.26.4
pandas
Pillow
pytest==8.3.3
pytorch_lightning==2.3.0
scikit_learn
scipy==1.7.3
tensorflow
termcolor==2.5.0
threadpoolctl==3.5.0
timm
torch
jupyter
opencv-python
transformers
torchvision
tqdm
vol==0.1.1
wandb
celer
torchsummary
cblearn
ttach
einops
pymf